fix(lint): split cli.rs family under the 500-line gate
